Taxi Demand Forecasting Pipeline using BigQuery ML & Python/requirements.txt
google-cloud-bigquery
google-cloud-bigquery-storage
pandas
numpy
matplotlib
plotly
python-dotenv
db-dtypes